Move Version module into lock version
This was only used in one place, and was non obvious.
Also create bang version to do the saving at the same time to try clean
up code.
